walrep: Use regular libpq protocol
We were using the GPDB internal protocol for streaming replication
connections. This can be traced back a long way (circa 2009), introduced
from a side effect in a historical commit. The change got carried over
implicitly when walrep was introduced.

There is no reason as to why this is necessary, as the internal protocol
is meant for special QD<->QE libpq interaction (fts, dispatch etc). For
walsender<->walreceiver connections, this is not necessary.

Note: Some fixes to tests in src/test/walrep were needed.
The generate_ao* tests were failing because of this type of HBA error:

ERROR:  could not connect to the primary server: FATAL:  no pg_hba.conf
entry for replication connection from host "[local]", user "pivotal",
SSL off (gplibpq.c:363)

We fixed this by adding 'host=localhost' to the connection string in
test_xlog_ao_wrapper(), since we have HBA entries for localhost.

This ERROR was actually causing a segfault as we weren't checking the
return value of walrcv_connect(). Now, we do so. Also, we take the
opportunity to make other walrcv() calls safe.

There is no reason as to why this is necessary, as the internal protocol
is meant for special QD<->QE libpq interaction (fts, dispatch etc). For
walsender<->walreceiver connections, this is not necessary.