gpstop does not work if previous gpstop was interrupted
Issue:
When a gpstop process with 'smart' option is interrupted, the subsequent
gpstop command becomes stuck with an error message stating "the database
system is shutting down". This error message also appears when attempting
to connect to the database using psql and other utilities.
The problem is that the error message does not provide clear instructions
on how to stop the database, leading to confusion and an inability to
perform a graceful shutdown. Gpstop will not work at this point and the
only way to stop the database is to manually kill the coordinator processes.

RCA
The issue occurs due to the interruption of the gpstop process, which
leaves the database system in an incomplete shutdown state with unfinished
coordinator processes. Also, this will not allow any further operations
on the database until the shutdown is complete.

Solution:
To address this issue, the gpstop code will be modified to handle the
scenario where the second attempt of gpstop displays a warning stating
"The database is currently in the process of shutting down.".
After this warning, if the gpstop is not running with the option '-u',
the code will prompt the user to choose between a "Fast" or "Immediate"
shutdown option. Depending on the user's selection, the code will proceed
as follows:

1. Terminate all ongoing unfinished coordinator processes from the
previous gpstop.
2. Start the cluster in coordinator-only mode to gather segment
information.
3. Perform a graceful stop of the cluster, completing the unfinished
shutdown process.