Removing stale replication entries from pg_hba.conf
In a multi host cluster, when we move one mirror segment to new host we are not removing the replication entry for old host from the primary's pg_hba.conf. This will result the primary's walsender to connect back to the old mirror's walreciever once the new host is unreachable and old host is available. Because of this, in gp_segment_configuration, the mirrors that belong to new host will be marked as up which is incorrect since new host is still unreachable. This commit include the code changes for removing deprecated replication entries from primary's pg_hba.conf while moving mirrors to new host.